requirements.txt
numpy==1.26.0
scipy==1.11.1
Pillow==10.0.0
keras==2.12.0